# public/swagger/api.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a api document of the sample node.js server framework.",
        "version": "1.0.0",
        "title": "RESTful API Document",
        "termsOfService": "",
        "contact": {
            "email": "super_xp@126.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "host": "",
    "basePath": "/",
...
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/api/utils/captcha": {
            "get": {
                "tags": [
                    "utils"
                ],
                "summary": "获取验证码",
                "description": "获取验证码",
                "operationId": "captcha",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "width",
                        "in": "query",
                        "description": "宽",
                        "required": true,
                        "type": "string",
                        "default": "80"
                    },
                    {
                        "name": "height",
                        "in": "query",
                        "description": "高",
                        "required": true,
                        "type": "string",
                        "default": "30"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        },
                        "headers": {
                            "X-Rate-Limit": {
                                "type": "integer",
                                "format": "int32",
                                "description": "calls per hour allowed by the user"
                            },
                            "X-Expires-After": {
                                "type": "string",
                                "format": "date-time",
                                "description": "date in UTC when token expires"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid username/password supplied"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "api_key": {
            "type": "apiKey",
            "name": "authorization",
            "in": "header"
        }
    },
...
}